--- backend/kirikiri/src/main/java/co/kirikiri/service/dto/member/GenderType.java
@@ -1,5 +1,0 @@
-package co.kirikiri.service.dto.member;
-
-public enum GenderType {
-    MALE, FEMALE
-}
