+++ backend/kirikiri/src/main/java/co/kirikiri/service/dto/member/request/GenderType.java
@@ -1,0 +1,5 @@
+package co.kirikiri.service.dto.member.request;
+
+public enum GenderType {
+    MALE, FEMALE
+}
