+++ backend/kirikiri/src/main/java/co/kirikiri/service/dto/member/response/MemberResponse.java
@@ -1,4 +1,4 @@
-package co.kirikiri.service.dto.member;
+package co.kirikiri.service.dto.member.response;
 
 public record MemberResponse(
         long id,
