+++ backend/kirikiri/src/main/java/co/kirikiri/service/dto/roadmap/RoadmapSaveDto.java
@@ -1,5 +1,6 @@
 package co.kirikiri.service.dto.roadmap;
 
+import co.kirikiri.service.dto.roadmap.request.RoadmapDifficultyType;
 import java.util.List;
 
 public record RoadmapSaveDto(
