+++ backend/kirikiri/src/main/java/co/kirikiri/service/dto/roadmap/request/RoadmapDifficultyType.java
@@ -1,4 +1,4 @@
-package co.kirikiri.service.dto.roadmap;
+package co.kirikiri.service.dto.roadmap.request;
 
 public enum RoadmapDifficultyType {
