+++ backend/kirikiri/src/main/java/co/kirikiri/service/dto/roadmap/request/RoadmapFilterTypeRequest.java
@@ -1,13 +1,13 @@
-package co.kirikiri.service.dto.roadmap;
+package co.kirikiri.service.dto.roadmap.request;
 
-public enum RoadmapFilterTypeDto {
+public enum RoadmapFilterTypeRequest {
     LATEST("최신순"),
     GOAL_ROOM_COUNT("생성된 골룸이 많은순"),
     PARTICIPANT_COUNT("참여 중인 인원순");
 
     private final String description;
 
-    RoadmapFilterTypeDto(final String description) {
+    RoadmapFilterTypeRequest(final String description) {
         this.description = description;
     }
 }
