+++ backend/kirikiri/src/test/java/co/kirikiri/integration/MemberIntegrationTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import co.kirikiri.integration.helper.IntegrationTest;
 import co.kirikiri.service.dto.ErrorResponse;
-import co.kirikiri.service.dto.member.GenderType;
+import co.kirikiri.service.dto.member.request.GenderType;
 import co.kirikiri.service.dto.member.request.MemberJoinRequest;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
