+++ backend/kirikiri/src/test/java/co/kirikiri/service/RoadmapServiceTest.java
@@ -26,14 +26,14 @@ import co.kirikiri.persistence.roadmap.RoadmapContentRepository;
 import co.kirikiri.persistence.roadmap.RoadmapRepository;
 import co.kirikiri.service.dto.CustomPageRequest;
 import co.kirikiri.service.dto.PageResponse;
-import co.kirikiri.service.dto.member.MemberResponse;
-import co.kirikiri.service.dto.roadmap.RoadmapCategoryResponse;
-import co.kirikiri.service.dto.roadmap.RoadmapContentResponse;
-import co.kirikiri.service.dto.roadmap.RoadmapDifficultyType;
-import co.kirikiri.service.dto.roadmap.RoadmapFilterTypeDto;
-import co.kirikiri.service.dto.roadmap.RoadmapNodeSaveRequest;
-import co.kirikiri.service.dto.roadmap.RoadmapResponse;
-import co.kirikiri.service.dto.roadmap.RoadmapSaveRequest;
+import co.kirikiri.service.dto.member.response.MemberResponse;
+import co.kirikiri.service.dto.roadmap.request.RoadmapDifficultyType;
+import co.kirikiri.service.dto.roadmap.request.RoadmapFilterTypeRequest;
+import co.kirikiri.service.dto.roadmap.request.RoadmapNodeSaveRequest;
+import co.kirikiri.service.dto.roadmap.request.RoadmapSaveRequest;
+import co.kirikiri.service.dto.roadmap.response.RoadmapCategoryResponse;
+import co.kirikiri.service.dto.roadmap.response.RoadmapContentResponse;
+import co.kirikiri.service.dto.roadmap.response.RoadmapResponse;
 import java.time.LocalDate;
 import java.util.Collections;
 import java.util.List;
@@ -168,7 +168,7 @@ class RoadmapServiceTest {
                 .thenReturn(Optional.empty());
 
         final Long categoryId = 1L;
-        final RoadmapFilterTypeDto filterType = RoadmapFilterTypeDto.LATEST;
+        final RoadmapFilterTypeRequest filterType = RoadmapFilterTypeRequest.LATEST;
         final CustomPageRequest pageRequest = new CustomPageRequest(1, 10);
 
         // expected
@@ -188,7 +188,7 @@ class RoadmapServiceTest {
                 .thenReturn(roadmapPages);
 
         final Long categoryId = 1L;
-        final RoadmapFilterTypeDto filterType = null;
+        final RoadmapFilterTypeRequest filterType = null;
         final CustomPageRequest pageRequest = new CustomPageRequest(1, 10);
 
         // when
@@ -219,7 +219,7 @@ class RoadmapServiceTest {
                 .thenReturn(roadmapPages);
 
         final Long categoryId = null;
-        final RoadmapFilterTypeDto filterType = RoadmapFilterTypeDto.LATEST;
+        final RoadmapFilterTypeRequest filterType = RoadmapFilterTypeRequest.LATEST;
         final CustomPageRequest pageRequest = new CustomPageRequest(1, 10);
 
         // when
@@ -252,7 +252,7 @@ class RoadmapServiceTest {
                 .thenReturn(roadmapPages);
 
         final Long categoryId = 1L;
-        final RoadmapFilterTypeDto filterType = RoadmapFilterTypeDto.LATEST;
+        final RoadmapFilterTypeRequest filterType = RoadmapFilterTypeRequest.LATEST;
         final CustomPageRequest pageRequest = new CustomPageRequest(1, 10);
 
         // when
